из-за <React.StrictMode> в index.js два раза дергается редюсер в useMessages.js, убрал <React.StrictMode>
