+++ src/index.js
@@ -5,10 +5,15 @@ import App from './App';
 import reportWebVitals from './reportWebVitals';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
+/*
 root.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>
+);*/
+
+root.render(
+        <App />
 );
 
 // If you want to start measuring performance in your app, pass a function
